minor refactoring, fix unnnecessary scrollbar, fix recent items on web
