+++ public/index.html
@@ -22,7 +22,7 @@
     <title>Universal Data Tool</title>
     <style>
       body {
-        overflow-y: scroll;
+        overflow-y: hidden;
       }
     </style>
     <!-- Global site tag (gtag.js) - Google Analytics -->
