+++ src/components/SamplesView/index.js
